Here is a demo of the trained model for the EV forecasting per building
This code takes the dataset that has been created and uses that data to train the SVM, Random Forest and XGBoost Models. Then it see's which model is the best and uses that to get the predicted EV owner per building. Then we can compare it to see how close it is to the original predicted EV owners per building to see how close it is.